refactor: 캘린더 엔터티 필드 및 업데이트 메서드 정리, 파일 위치 변경
